correct schemas and topics
